setup async/await test suite
