+++ package.json
@@ -44,6 +44,7 @@
     "fastify-compress": "^0.8.1",
     "pre-commit": "^1.2.2",
     "proxyquire": "^2.1.0",
+    "semver": "^6.3.0",
     "simple-get": "^3.0.3",
     "snazzy": "^8.0.0",
     "standard": "^12.0.0",
