+++ test/static.test.js
@@ -548,75 +548,6 @@ t.test('sendFile', t => {
   })
 })
 
-t.test('sendFile with async/await', t => {
-  t.plan(2)
-
-  const pluginOptions = {
-    root: path.join(__dirname, '/static'),
-    prefix: '/static'
-  }
-  const fastify = Fastify()
-  fastify.register(fastifyStatic, pluginOptions)
-
-  fastify.get('/foo/bar', async function (req, reply) {
-    await reply.sendFile('/index.html')
-  })
-
-  fastify.listen(0, err => {
-    t.error(err)
-
-    fastify.server.unref()
-
-    t.test('reply.sendFile()', t => {
-      t.plan(3 + GENERIC_RESPONSE_CHECK_COUNT)
-      simple.concat({
-        method: 'GET',
-        url: 'http://localhost:' + fastify.server.address().port + '/foo/bar',
-        followRedirect: false
-      }, (err, response, body) => {
-        t.error(err)
-        t.strictEqual(response.statusCode, 200)
-        t.strictEqual(body.toString(), indexContent)
-        genericResponseChecks(t, response)
-      })
-    })
-  })
-})
-
-t.test('sendFile error with async/await', t => {
-  t.plan(2)
-
-  const pluginOptions = {
-    root: path.join(__dirname, '/static'),
-    prefix: '/static'
-  }
-  const fastify = Fastify()
-  fastify.register(fastifyStatic, pluginOptions)
-
-  fastify.get('/foo/bar', async function (req, reply) {
-    await reply.sendFile('/index-not-exists.html')
-  })
-
-  fastify.listen(0, err => {
-    t.error(err)
-
-    fastify.server.unref()
-
-    t.test('reply.sendFile()', t => {
-      t.plan(3)
-      simple.concat({
-        method: 'GET',
-        url: 'http://localhost:' + fastify.server.address().port + '/foo/bar',
-        followRedirect: false
-      }, (err, response, body) => {
-        t.error(err)
-        t.strictEqual(response.statusCode, 404)
-        t.strictEqual(body.toString(), '{"message":"Route GET:/foo/bar not found","error":"Not Found","statusCode":404}')
-      })
-    })
-  })
-})
-
 t.test('sendFile disabled', t => {
   t.plan(2)
